Fix table layout
